requirements.txt
numpy==1.24.3
pandas==2.1.4
python-dotenv==1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-ollama>=0.2.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
unstructured==0.10.30
huggingface-hub
torch
transformers